:recycle: Extract other file
